+++ RELEASE_NOTES.md
@@ -1,2 +1,4 @@
 v0.86.17
 
+- Expose accountExists() API to client libraries
+